Move stream functionality from BaseProvider to new StreamProvider
