+++ src/utils.ts
@@ -1,4 +1,3 @@
-import { EventEmitter } from 'events';
 import { JsonRpcMiddleware, PendingJsonRpcResponse } from 'json-rpc-engine';
 import { ethErrors } from 'eth-rpc-errors';
 
@@ -57,35 +56,10 @@ export const getRpcPromiseCallback =
     }
   };
 
-/**
- * Logs a stream disconnection error. Emits an 'error' if given an
- * EventEmitter that has listeners for the 'error' event.
- *
- * @param log - The logging API to use.
- * @param remoteLabel - The label of the disconnected stream.
- * @param error - The associated error to log.
- * @param emitter - The logging API to use.
- */
-export function logStreamDisconnectWarning(
-  log: ConsoleLike,
-  remoteLabel: string,
-  error: Error,
-  emitter: EventEmitter,
-): void {
-  let warningMsg = `MetaMask: Lost connection to "${remoteLabel}".`;
-  if (error?.stack) {
-    warningMsg += `\n${error.stack}`;
-  }
-  log.warn(warningMsg);
-  if (emitter && emitter.listenerCount('error') > 0) {
-    emitter.emit('error', warningMsg);
-  }
-}
-
 export const NOOP = () => undefined;
 
 // constants
 
-export const EMITTED_NOTIFICATIONS = [
+export const EMITTED_NOTIFICATIONS = Object.freeze([
   'eth_subscription', // per eth-json-rpc-filters/subscriptionManager
-];
+]);
